More changes

+++ sdk/core/core/inc/az_span.h
@@ -92,7 +92,7 @@ AZ_NODISCARD AZ_INLINE az_span az_span_init(uint8_t * ptr, int32_t length, int32
  *
  */
 AZ_NODISCARD AZ_INLINE az_span az_span_from_str(char * str) {
-  int32_t length = strlen(str);
+  int32_t length = (int32_t)strlen(str);
   return az_span_init((uint8_t *)str, length, length);
 }
